interp: a reference is a RefHandle, not a function index
Table entries, globals, and slots held a bare u32 function index with
u32::MAX as the null funcref, so externref and anyref had no
representation and every non-funcref table was refused.

The substrate already had one. A plain `RefHandle`'s payload IS the
function index (`RefHandle::new(func_idx)` in jit/instance.rs), and
extern/host refs ride in its tag bits -- so this is a widening, not a new
representation: funcrefs keep the same bits and the rest arrives with
them. Slots normalize null to u64::MAX rather than storing `usize::MAX`
verbatim, so a reference means the same thing in a slot on a 32-bit
target as on a 64-bit one.

The host boundary converts `Value::Ref` with the same two functions the
engine uses internally, so the interpreter and the JIT hand the identical
handle across.

The generated call_indirect handlers read table 0 directly and had the
4-byte stride, the 32-bit load, and the -1 null test baked into the
assembly on arm64, x86_64, and riscv. All three now read 8-byte slots.
arm64 and x86_64 test the high half, which rejects a tagged handle as
well as a null; riscv keeps its all-ones compare. arm32 routes
call_indirect to the slow path and was unaffected.

A grep for `table0` in interp_gen finds nothing -- the handlers reach it
by struct offset (#112), not by name. The unit test caught it: index 1
dispatched to function 0, because a 4-byte stride into 8-byte slots lands
on the zero high half of entry 0.

interp spectest 102 -> 111. JIT 257/257 unchanged; 411 tests.

+++ sf-nano-core/src/vm/interpreter/predecode.rs
@@ -34,7 +34,7 @@ const NO_DEF: u32 = u32::MAX;
 /// Branch-target placeholder while the target's `end` is still ahead.
 const FIXUP: u64 = u64::MAX;
 /// Null funcref representation (function indices are table/ref values).
-pub(super) const NULL_FUNCREF: u64 = u32::MAX as u64;
+pub(super) const NULL_FUNCREF: u64 = u64::MAX;
 
 pub(crate) struct PredecodedFunction {
     pub code: Vec<Instr>,
